% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{eigen_mean_R}
\alias{eigen_mean_R}
\title{Compute mean using Eigen mean reduction method.}
\usage{
eigen_mean_R(x.vec)
}
\arguments{
\item{x.vec}{vector of data which will be coerced using as.double}
}
\value{
mean of data
}
\description{
Compute mean using Eigen mean reduction method.
}
\examples{
eigen_mean_R(c(1,2))
}
