% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{my_mean_R}
\alias{my_mean_R}
\title{Compute mean using pure C}
\usage{
my_mean_R(x.vec)
}
\arguments{
\item{x.vec}{vector with at least 1 element, which is coerced to integer}
}
\value{
mean, numeric/double scalar
}
\description{
Compute mean using pure C
}
\examples{
my_mean_R(c(1,2))
}
